move timer metrics to a standalone package
