+++ metrics/BUILD.bazel
@@ -19,7 +19,6 @@ go_library(
         "sli.go",
         "stats.go",
         "telemetry.go",
-        "timer.go",
         "topsql.go",
         "ttl.go",
         "wrapper.go",
@@ -28,6 +27,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//parser/terror",
+        "//timer/metrics",
         "//util/logutil",
         "//util/mathutil",
         "@com_github_pingcap_errors//:errors",
