+++ timer/runtime/BUILD.bazel
@@ -10,8 +10,8 @@ go_library(
     importpath = "github.com/pingcap/tidb/timer/runtime",
     visibility = ["//visibility:public"],
     deps = [
-        "//metrics",
         "//timer/api",
+        "//timer/metrics",
         "//util/logutil",
         "@com_github_google_uuid//:uuid",
         "@com_github_pingcap_errors//:errors",
@@ -35,9 +35,9 @@ go_test(
     race = "on",
     shard_count = 19,
     deps = [
-        "//metrics",
         "//testkit/testsetup",
         "//timer/api",
+        "//timer/metrics",
         "//util/mock",
         "@com_github_google_uuid//:uuid",
         "@com_github_pingcap_errors//:errors",
